+++ DoRunDoRun/Sources/Domain/Entity/Goal/GoalOption.swift
@@ -5,13 +5,14 @@
 //  Created by Jaehui Yu on 9/24/25.
 //
 
-import Foundation
-
 struct GoalOption {
     let type: GoalOptionType
-    let distance: [Int]         // m 단위
-    let duration: [Int]         // 분 단위
-    let pace: [Int]             // 초/킬로미터 단위 (서버가 분 단위를 기대하면 변환해서 보냄)
+    /// m 단위
+    let distance: [Int]
+    /// 분 단위
+    let duration: [Int]
+    // 초/킬로미터 단위
+    let pace: [Int]
 }
 
 enum GoalOptionType: String {
