+++ DoRunDoRun/Sources/Scene/GoalSettingFlow/RunningLevelCheck/RunningLevelCheckViewController.swift
@@ -169,7 +169,7 @@ final class RunningLevelCheckViewController: UIViewController {
     // MARK: Actions
     
     @objc private func didTapNext() {
-        router?.routeToGoalSetting()
+        router?.routeToGoalOptionSelect()
     }
 }
 
@@ -205,91 +205,5 @@ extension RunningLevelCheckViewController: RunningLevelCheckDisplayLogic {
             IndexPath(row: viewModel.selectedIndex, section: 0)
         ]
         tableView.reloadRows(at: indexPaths, with: .none)
-    }
-}
-
-// MARK: - RunningLevel
-
-final class RunningLevelCell: UITableViewCell {
-    static let identifier = "RunningLevelCell"
-    
-    // MARK: UI
-    
-    private let iconView = UIImageView()
-    private let titleLabel = UILabel()
-    private let subtitleLabel = UILabel()
-    
-    private let container: UIView = {
-        let view = UIView()
-        view.layer.cornerRadius = 12
-        view.layer.borderWidth = 1
-        return view
-    }()
-    
-    // MARK: Object lifecycle
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        setupView()
-    }
-    
-    @available(*, unavailable)
-    required init?(coder: NSCoder) { fatalError("init(coder:) has not been implemented") }
-    
-    // MARK: Setup
-    
-    private func setupView() {
-        selectionStyle = .none
-        
-        contentView.addSubview(container)
-        container.translatesAutoresizingMaskIntoConstraints = false
-        
-        [iconView, titleLabel, subtitleLabel].forEach {
-            container.addSubview($0)
-            $0.translatesAutoresizingMaskIntoConstraints = false
-        }
-        
-        NSLayoutConstraint.activate([
-            container.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 8),
-            container.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -8),
-            container.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 20),
-            container.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -20),
-            
-            iconView.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: 16),
-            iconView.centerYAnchor.constraint(equalTo: container.centerYAnchor),
-            iconView.widthAnchor.constraint(equalToConstant: 46),
-            iconView.heightAnchor.constraint(equalToConstant: 46),
-            
-            titleLabel.leadingAnchor.constraint(equalTo: iconView.trailingAnchor, constant: 12),
-            titleLabel.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: -16),
-            titleLabel.topAnchor.constraint(equalTo: container.topAnchor, constant: 16),
-            
-            subtitleLabel.leadingAnchor.constraint(equalTo: titleLabel.leadingAnchor),
-            subtitleLabel.trailingAnchor.constraint(equalTo: titleLabel.trailingAnchor),
-            subtitleLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 4),
-            subtitleLabel.bottomAnchor.constraint(lessThanOrEqualTo: container.bottomAnchor, constant: -16)
-        ])
-    }
-    
-    // MARK: Configure
-    
-    func configure(with runningLevel: DisplayedRunningLevel) {
-        iconView.image = UIImage(systemName: runningLevel.image)
-
-        titleLabel.attributedText = .withLetterSpacing(
-            text: runningLevel.title,
-            font: .pretendard(size: 18, weight: .bold),
-            px: -0.2,
-            color: .init(hex: 0x232529)
-        )
-        
-        subtitleLabel.attributedText = .withLetterSpacing(
-            text: runningLevel.subtitle,
-            font: .pretendard(size: 14, weight: .regular),
-            px: -0.2,
-            color: .init(hex: 0x585D64)
-        )
-        
-        container.layer.borderColor = runningLevel.isSelected ? UIColor(hex: 0x3E4FFF).cgColor : UIColor(hex: 0xDFE4EC).cgColor
     }
 }
